Cleanup of membership list

+++ internal/encoding/member.go
@@ -10,7 +10,7 @@ type Member struct {
 
 	// IncarnationNumber is the incarnation the member gave about itself. It is monotonically increasing each time
 	// somebody suspects the member. Only the member itself is allowed to increase the incarnation.
-	IncarnationNumber int
+	IncarnationNumber uint16
 
 	// SuspicionPeriodCounter is the number of protocol periods the member is in a suspicion state. This is useful when
 	// deciding about declaring a member as faulty when it was under suspicion long enough.
